code/requirements.txt
# Core Framework
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-ollama>=0.1.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4
sqlalchemy>=2.0.0

# Graph Database
neo4j>=5.15.0
# falkordb>=1.0.0  # Alternative graph DB

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0  # Fast DataFrame operations

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
selenium>=4.16.0  # For dynamic content
youtube-transcript-api>=0.6.0 # For retrieving YouTube captions
scrapetube>=2.5.0 # For fetching video lists without API key

# Document Processing
# unstract>=0.1.0  # Not yet available on PyPI - use alternative
unstructured>=0.10.0  # Alternative document processing
pypdf>=3.17.0
python-docx>=1.1.0

# NLP & Embeddings
sentence-transformers>=2.2.0
spacy>=3.7.0
transformers>=4.36.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.15.0  # Map visualization

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.28.0
